requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
streamlit>=1.28.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langgraph>=0.0.40
zai-sdk>=0.2.2
chromadb>=0.4.0
rank-bm25>=0.2.2
PyMuPDF>=1.23.0
python-dotenv>=1.0.0
requests>=2.31.0